# project_back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestion des Exercices",
        "version": "1.0"
    },
    "paths": {
        "/exercices": {
            "get": {
                "tags": [
                    "Exercices"
                ],
                "summary": "Liste tous les exercices",
                "operationId": "e558c39539d68edb9fdec37196f9aebf",
                "responses": {
                    "200": {
                        "description": "Liste des exercices récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Exercice"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Exercices"
                ],
                "summary": "Créer un nouvel exercice",
                "operationId": "c221ea923daadc2bbcdda1dd62d9a457",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Pompes"
                                    },
                                    "advice": {
                                        "type": "string",
                                        "example": "Gardez le dos droit"
                                    },
                                    "series": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "repetitions": {
                                        "type": "integer",
                                        "example": 12
                                    },
                                    "id_training": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Exercice créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Exercice"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Séance non trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Séance non trouvée"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/exercices/{id}": {
            "get": {
                "tags": [
                    "Exercices"
                ],
                "summary": "Affiche les détails d'un exercice",
                "operationId": "5941d5da0480f89b22368448de8b23b6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'exercice",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de l'exercice récupérés",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Exercice"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Exercice non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Exercice non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Exercices"
                ],
                "summary": "Mettre à jour un exercice",
                "operationId": "c9b41cb1bb4fcbb1ab615203aa90e53f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'exercice",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Pompes"
                                    },
                                    "advice": {
                                        "type": "string",
                                        "example": "Gardez le dos droit"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Exercice mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Exercice"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Exercice non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Exercice non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Exercices"
                ],
                "summary": "Supprimer un exercice",
                "operationId": "dda7c2d1fdbfa0a4b9b93e4a64c86fdb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'exercice",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exercice supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Exercice supprimé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Exercice non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Exercice non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Exercice": {
                "title": "Exercice",
                "description": "Un modèle représentant un exercice d'entraînement",
                "properties": {
                    "id": {
                        "description": "ID de l'exercice",
                        "type": "integer",
                        "example": 1
                    },
                    "nom": {
                        "description": "Nom de l'exercice",
                        "type": "string",
                        "example": "Pompes"
                    },
                    "description": {
                        "description": "Description de l'exercice",
                        "type": "string",
                        "example": "Exercice de renforcement musculaire"
                    },
                    "series": {
                        "description": "Nombre de séries",
                        "type": "integer",
                        "example": 3
                    },
                    "repetitions": {
                        "description": "Nombre de répétitions",
                        "type": "integer",
                        "example": 12
                    },
                    "training_id": {
                        "description": "ID de la séance associée à cet exercice",
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "description": "Date de création de l'exercice",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date de dernière mise à jour de l'exercice",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}